Extract stencil pass to its own file
